backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
tensorflow==2.20.0
Pillow==10.0.1
numpy>=1.26.0
python-dotenv==1.0.0